Version 2.0 Now features clicking
Redid a lot of the structuring. The process of making new quizzes is now much more straightforward. Edit the "quiz_#.js" line in the "quiz_#.html" file and fill the information in the JS file. Other than that, all that is needed is the images and updating "index.html". The "quiz_#_map.html" files still need to be done individually.