Tool used in preparing "quiz_##_Maps.html" files
I made this tool just to add text over images and see modifications to the text's CSS without needing to refresh the page after each modification. After all the text for a map is placed with the desired options, a list of the CSS attributes is provided at the bottom of the screen.